Changed K values and Gyro retrieval
Gyro and acceleration are now retrieved by using the raw values
from the chip as opposed to the DMP values.

Also, note that DMP is being left in the code as it appears to
prevent the library from crashing on the arduino for whatever
reason.  Potentially this could be due to the DMP's handling of
buffer overflow.

Things to consider:
 - adding the data aquisition to main loop
 - adding data calculation to main loop
 - motor control call to main loop

As opposed to having the motor coltron class do the calls for
data retrieval and calculation...perhaps this would add the
main loop's timing sequence into play and might add accuracy to
the return of the Kalman filter.